Ejercicios Part 1 - Terminados hasta el 10

+++ Tema1/src/parte1/Ejercicio4.java
@@ -13,10 +13,10 @@ public class Ejercicio4 {
 		Scanner sc = new Scanner(System.in);
 
 		// Creamos la variable nota1 para almacenar la primera nota.
-		double nota1;
+		int nota1;
 
 		// Creamos la variable nota2 para almacenar la segunda nota.
-		double nota2;
+		int nota2;
 
 		// Creamos la variable notaMedia para almacenar el resultado de la media de
 		// estos dos números.
@@ -25,16 +25,16 @@ public class Ejercicio4 {
 		// Le pedimos al usuario que introduzca la primera nota.
 		System.out.print("Introduce la primera nota --> ");
 		// Leemos la primera nota del teclado.
-		nota1 = sc.nextDouble();
+		nota1 = sc.nextInt();
 
 		// Le pedimos al usuario que introduzca la segunda nota.
 		System.out.print("Introduce la primera nota --> ");
 		// Leemos la segunda nota del teclado.
-		nota2 = sc.nextDouble();
+		nota2 = sc.nextInt();
 
 		// A la variable nota media le asignamos la ecuación para calcular la nota media
 		// entre estas dos notas.
-		notaMedia = (nota1 + nota2) / 2;
+		notaMedia = (double) (nota1 + nota2) / 2;
 
 		// Mostramos la nota media
 		System.out.println("La nota media de estas dos notas es: " + notaMedia);
